+++ src/events/sl-after-hide.ts
@@ -1,0 +1,9 @@
+type SlAfterHideEvent = CustomEvent<Record<PropertyKey, never>>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'sl-after-hide': SlAfterHideEvent;
+  }
+}
+
+export default SlAfterHideEvent;