+++ src/events/sl-error.ts
@@ -1,0 +1,9 @@
+type SlErrorEvent = CustomEvent<{ status?: number }>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'sl-error': SlErrorEvent;
+  }
+}
+
+export default SlErrorEvent;