+++ src/events/sl-finish.ts
@@ -1,0 +1,9 @@
+type SlFinishEvent = CustomEvent<Record<PropertyKey, never>>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'sl-finish': SlFinishEvent;
+  }
+}
+
+export default SlFinishEvent;