+++ src/events/sl-invalid.ts
@@ -1,0 +1,9 @@
+type SlInvalidEvent = CustomEvent<Record<PropertyKey, never>>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'sl-invalid': SlInvalidEvent;
+  }
+}
+
+export default SlInvalidEvent;