+++ src/events/sl-mutation.ts
@@ -1,0 +1,9 @@
+type SlMutationEvent = CustomEvent<{ mutationList: MutationRecord[] }>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'sl-mutation': SlMutationEvent;
+  }
+}
+
+export default SlMutationEvent;