+++ src/shoelace.ts
@@ -10,6 +10,8 @@ export { default as SlButton } from './components/button/button';
 export { default as SlButtonGroup } from './components/button-group/button-group';
 export { default as SlCalendar } from './components/calendar/calendar';
 export { default as SlCard } from './components/card/card';
+export { default as SlCarousel } from './components/carousel/carousel';
+export { default as SlCarouselItem } from './components/carousel-item/carousel-item';
 export { default as SlCheckbox } from './components/checkbox/checkbox';
 export { default as SlColorPicker } from './components/color-picker/color-picker';
 export { default as SlDetails } from './components/details/details';
@@ -61,3 +63,6 @@ export { default as SlOption } from './components/option/option';
 export * from './utilities/animation';
 export * from './utilities/base-path';
 export * from './utilities/icon-library';
+
+// Events
+export * from './events/events';