+++ src/translations/en.ts
@@ -9,16 +9,19 @@ const translation: Translation = {
   clearEntry: 'Clear entry',
   close: 'Close',
   copy: 'Copy',
+  currentValue: 'Current value',
+  goToSlide: (slide, count) => `Go to slide ${slide} of ${count}`,
+  hidePassword: 'Hide password',
+  loading: 'Loading',
+  nextMonth: 'Next month',
+  nextSlide: 'Next slide',
   numOptionsSelected: num => {
     if (num === 0) return 'No options selected';
     if (num === 1) return '1 option selected';
     return `${num} options selected`;
   },
-  currentValue: 'Current value',
-  nextMonth: 'Next Month',
-  previousMonth: 'Previous Month',
-  hidePassword: 'Hide password',
-  loading: 'Loading',
+  previousMonth: 'Previous month',
+  previousSlide: 'Previous slide',
   progress: 'Progress',
   remove: 'Remove',
   resize: 'Resize',