+++ src/translations/pt.ts
@@ -9,14 +9,17 @@ const translation: Translation = {
   clearEntry: 'Limpar entrada',
   close: 'Fechar',
   copy: 'Copiar',
+  currentValue: 'Valor atual',
+  goToSlide: (slide, count) => `Vá para o slide ${slide} de ${count}`,
+  hidePassword: 'Esconder a senha',
+  loading: 'Carregando',
+  nextSlide: 'Próximo slide',
   numOptionsSelected: num => {
     if (num === 0) return 'Nenhuma opção selecionada';
     if (num === 1) return '1 opção selecionada';
     return `${num} opções selecionadas`;
   },
-  currentValue: 'Valor atual',
-  hidePassword: 'Esconder a senha',
-  loading: 'Carregando',
+  previousSlide: 'Slide anterior',
   progress: 'Progresso',
   remove: 'Remover',
   resize: 'Mudar o tamanho',